# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"ingrid-bicudo","description":"bootcamp-w1-g7-mlb-frescos","contact":{"name":"ingrid-bicudo","email":"ingrid.bicudo@mercadolivre.com"},"version":"last"},"servers":[{"url":"http://localhost:8080","description":"Scope local"}],"paths":{"/api/v1/fresh-products/section":{"put":{"tags":["section-controller"],"description":"Update an existent inbound order and return a list with the batches updated","operationId":"updateSection","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionUpdateDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Section"}}}}}}},"/api/v1/fresh-products/orders":{"get":{"tags":["order-controller"],"description":"Get a list of products of an a existing order","operationId":"listProductFromOrder","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrder"}}}}}}},"put":{"tags":["order-controller"],"description":"Update an existing order for a client and return the price of the order","operationId":"craeteOrder","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOrderResponseDTO"}}}}}},"post":{"tags":["order-controller"],"description":"Create an Order for a client and return the price of the order","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOrderResponseDTO"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["inbound-order-controller"],"description":"Update an existent inbound order and return a list with the batches updated","operationId":"updateInboundOrder","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInboundOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockDTO"}}}}}},"post":{"tags":["inbound-order-controller"],"description":"Create an inbound order and return a list with the batches created","operationId":"createInboundOrder","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"The object with contains the new inbound order and their products batches","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInboundOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockDTO"}}}}}}},"/login":{"post":{"tags":["session-controller"],"description":"Returns a valid token with id and the role of the authenticated user for authorizations","operationId":"login","requestBody":{"description":"The user credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}}},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/warehouse":{"get":{"tags":["warehouse-controller"],"description":"Returns a list with the warehouse and the corresponding quantity of a specific product","operationId":"getProductsInAllWarehouses","parameters":[{"name":"productId","in":"query","description":"The product id to search","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWarehouseDTO"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["inbound-order-controller"],"description":"Returns a location of a specified product in the logged supervisor warehouse","operationId":"listProductBatchStock","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}},{"name":"productId","in":"query","description":"The product id to be listed","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"sortParam","in":"query","description":"The order to filter by. Can be 'F' for due date or 'C' for amount ","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductBatchStockDTO"}}}}}}},"/api/v1/fresh-products/due-date":{"get":{"tags":["batch-controller"],"description":"Returns a list with the batches, located in the supervisor's warehouse, that have products with due date less than the actual date + the number os days specified in the quantityDays param.","operationId":"getExpiringProductsDTO","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}},{"name":"quantityDays","in":"query","description":"The day amount to plus the actual date for filter the products by their due dates","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpiringProductsDTO"}}}}}}},"/api/v1/fresh-products/due-date/list":{"get":{"tags":["batch-controller"],"description":"A list filtered by product category with the batches, located in the supervisor's warehouse, that have products with due date less than the actual date + the number os days specified in the quantityDays param. The list can be sorted by due date in ascending or descending order passed by the typeOrder parameter.","operationId":"getExpiringProductsDTO_1","parameters":[{"name":"Authorization","in":"header","description":"The token generated in the authentication","required":true,"schema":{"type":"string"}},{"name":"quantityDays","in":"query","description":"The day amount to plus the actual date for filter the products by their due dates","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","description":"The category to filter by","required":true,"schema":{"type":"string"}},{"name":"typeOrder","in":"query","description":"The order to sort by. Can be 'asc' or 'desc'","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpiringProductsDTO"}}}}}}},"/api/v1/fresh-products/":{"get":{"tags":["products-controller"],"operationId":"getMethodName","parameters":[{"name":"filter","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductListResponseDTO"}}}}}}}},"components":{"schemas":{"ValidationError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"ApiError":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"SectionUpdateDTO":{"required":["capacity","category","code","warehouse"],"type":"object","properties":{"code":{"type":"string"},"category":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"warehouse":{"$ref":"#/components/schemas/WarehouseDTO"}}},"WarehouseDTO":{"required":["code","name"],"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}},"Batch":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"},"inboundOrder":{"$ref":"#/components/schemas/InboundOrder"},"product":{"$ref":"#/components/schemas/Product"}}},"InboundOrder":{"type":"object","properties":{"orderNumber":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/Section"},"batchStock":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Batch"}},"supervisor":{"$ref":"#/components/schemas/Supervisor"}}},"Product":{"required":["name"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"category":{"type":"string"},"seller":{"$ref":"#/components/schemas/Seller"},"price":{"type":"number","format":"double"}}},"Section":{"type":"object","properties":{"code":{"type":"string"},"category":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"warehouse":{"$ref":"#/components/schemas/Warehouse"},"inboundOrder":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/InboundOrder"}}}},"Seller":{"required":["name"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"maxLength":45,"minLength":2,"type":"string"}}},"Supervisor":{"required":["name"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"maxLength":45,"minLength":2,"type":"string"},"warehouse":{"$ref":"#/components/schemas/Warehouse"}}},"Warehouse":{"required":["name"],"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}},"CreateOrderRequestDTO":{"type":"object","properties":{"order":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrderDTO"}}}},"ProductOrderDTO":{"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"quantity":{"type":"integer","format":"int32"}}},"CreateOrderResponseDTO":{"type":"object","properties":{"price":{"type":"number","format":"double"}}},"BatchDTO":{"required":["batchNumber","currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"productId":{"type":"string","format":"uuid"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderDTO":{"required":["batchStock","orderDate","orderNumber","section"],"type":"object","properties":{"orderNumber":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"SectionDTO":{"required":["sectionCode","warehouseCode"],"type":"object","properties":{"sectionCode":{"type":"string"},"warehouseCode":{"type":"string"}}},"UpdateInboundOrderDTO":{"type":"object","properties":{"inboundOrder":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"BatchStockDTO":{"required":["batchStock"],"type":"object","properties":{"batchStock":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"LoginRequestDTO":{"required":["password","userName"],"type":"object","properties":{"userName":{"type":"string"},"password":{"type":"string"}}},"LoginResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"token":{"type":"string"}}},"BatchWithoutNumberDTO":{"required":["currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"CreateInboundOrderDTO":{"required":["inboundOrder"],"type":"object","properties":{"inboundOrder":{"$ref":"#/components/schemas/InboundOrderWithoutOrderNumberDTO"}}},"InboundOrderWithoutOrderNumberDTO":{"required":["batchStock","orderDate","section"],"type":"object","properties":{"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchWithoutNumberDTO"}}}},"ProductWarehouseDTO":{"required":["productId","warehouses"],"type":"object","properties":{"productId":{"type":"string"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseQuantityDTO"}}}},"WarehouseQuantityDTO":{"required":["totalQuantity","warehouseCode"],"type":"object","properties":{"warehouseCode":{"type":"string"},"totalQuantity":{"type":"string"}}},"Order":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"price":{"type":"number","format":"double"},"detailOrder":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrder"}}}},"ProductOrder":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"order":{"$ref":"#/components/schemas/Order"},"product":{"$ref":"#/components/schemas/Product"},"quantity":{"type":"integer","format":"int32"}}},"BatchInfoDTO":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string","format":"date"}}},"ProductBatchStockDTO":{"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionDTO"},"productId":{"type":"string","format":"uuid"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchInfoDTO"}}}},"BatchStockWithDueDateDTO":{"required":["batchNumber","dueDate","productId","productTypeId","quantity"],"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"productId":{"type":"string"},"productTypeId":{"type":"string"},"dueDate":{"type":"string","format":"date"},"quantity":{"type":"integer","format":"int32"}}},"ExpiringProductsDTO":{"required":["batchStock"],"type":"object","properties":{"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockWithDueDateDTO"}}}},"ProductListResponseDTO":{"type":"object","properties":{"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}