# src/main/java/com/mercadolibre/ingrid_bicudo/dtos/LoginResponseDTO.java
package com.mercadolibre.ingrid_bicudo.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.UUID;

@AllArgsConstructor
@NoArgsConstructor
@Setter @Getter
public class LoginResponseDTO {
    private UUID id;
    private String token;
}